Commit de confirmacion de compra

+++ src/controllers/dao/manager/managerCart.mdb.js
@@ -1,6 +1,9 @@
 import { modelCart } from "../models/cart.model.js";
 import { modelProducts } from "../models/products.model.js";
 import { cartHistorial } from "../models/cart_historial.model.js";
+import { modelUsers } from "../models/users.model.js";
+import { config } from "../../config/config.js";
+import { transport_nodemailer } from "../../../services/utils/nodemailer.js";
 
 class ColectionManagerCart {
     static cart = [];
@@ -54,21 +57,37 @@ class ColectionManagerCart {
     static async endPurchase(_user_id) {
         try {
             const productCart = await modelCart.find({ _user_id: _user_id });
-            const cart = productCart.map(({ _product_id, quantity }) => ({ _product_id, quantity }));
-            const products = await Promise.all(cart.map(async (productFind) => {
-                return await modelProducts.findById(productFind._product_id)
-            }))
-            return await Promise.all(products.map(async ({ stock }) => {
-                cart.map(async ({ _product_id, quantity }) => {
-                    if (stock < quantity || quantity > stock) return console.error('El stock es mayor a la cantidad indicada en el carrito, la compra no puede ser procesada');
-                    const newStock = +stock - quantity;
-                    await modelProducts.updateOne({ _id: _product_id }, { $set: { stock: `${newStock}` } });
-                    await modelCart.findOneAndDelete({ _user_id: _user_id })
-                })
-            }));
+            const user = await modelUsers.findById(_user_id);
+            console.log(user)
+            if (productCart.length > 0) {
+                const cart = productCart.map(({ _product_id, quantity }) => ({ _product_id, quantity }));
+                const products = await Promise.all(cart.map(async (productFind) => {
+                    return await modelProducts.findById(productFind._product_id)
+                }))
+                return await Promise.all(products.map(async ({ stock }) => {
+                    cart.map(async ({ _product_id, quantity }) => {
+                        if (stock < quantity || quantity > stock) return console.error('El stock es mayor a la cantidad indicada en el carrito, la compra no puede ser procesada');
+                        const confirmation = await transport_nodemailer.sendMail({
+                            from: `Confirmacion de compra<${config.GMAIL_APP_USER}>`,
+                            to: user.email,
+                            subject: 'Confirmacion de compra',
+                            html: `
+                            <h1>¡Compra realizada con exito!</h1>
+                            
+                            `
+                        });
+                        console.log(confirmation);
+                        const newStock = +stock - quantity;
+                        await modelProducts.updateOne({ _id: _product_id }, { $set: { stock: `${newStock}` } });
+                        await modelCart.findOneAndDelete({ _user_id: _user_id })
+                    })
+                }));
+            } else {
+                console.error("Usuario no encontrado")
+            }
         }
         catch (err) {
-            console.error('Error al intentar terminar con la compra', err)
+            console.error('Error al intentar terminar con la compra', err.message)
         }
     }
     static async getProductCartById(id) {
